pyproject.toml
[project]
name = "cavepaint-net"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.9",
    "aiortc>=1.9",
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "mediapipe>=0.10.33",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pymongo>=4.16.0",
    "pyserial>=3.5",
    "pyzmq>=27.1.0",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
]
